% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta450k.m}
\alias{beta450k.m}
\title{DNA methylation profile (TCGA BRCA)}
\format{
A matrix with 391,633 rows (CpGs) and 10 variables (samples)
}
\usage{
data(beta450k.m)
}
\description{
An example data table that containing the DNA methylation beta
values from 10 bulk tissue samples of breast cancer from TCGA (5 normal vs. 5
tumor). Only used for demonstration purpose.
}
\keyword{datasets}
