% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmiR.R
\name{DoWilcox}
\alias{DoWilcox}
\title{Title Get differential analysis statistics}
\usage{
DoWilcox(data.m, pheno.v, case = "", control = "")
}
\arguments{
\item{data.m}{data matrix for average DNA methylation probes mapped miRNA
promoters or miRNA expression.}

\item{pheno.v}{The phenotype vector.}

\item{case}{Case group indicated in the phenotype vector.}

\item{control}{Control group indicated in the phenotype vector.}
}
\value{
data matrix for differential analysis statistics.
}
\description{
Title Get differential analysis statistics
}
\examples{
data(beta450k.m)
miRBeta.m <- probe2miR(beta=beta450k.m, array="450k", type="Mature", upstream=2000, downstream=2000, method="mean")
pheno.v  <- c(rep("Normal", 5), rep("Tumor", 5))
statM.df <- DoWilcox(miRBeta.m, pheno.v, case = "Tumor", control = "Normal")

}
