% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneBeta.R
\name{getGeneBeta}
\alias{getGeneBeta}
\title{Title Obtain beta value for each individual miRNA}
\usage{
getGeneBeta(pos, beta, upstream = 1500, downstream = 500, method = "mean")
}
\arguments{
\item{pos}{A vector consisting of the relative positions of the CpGs within
promoter region of a miRNA to its transcription start site, the names of
the elements in the vector should be the CpG ID of the Illumina Infinium
array, for example: "cg26257082".}

\item{beta}{The DNA methylation beta value matrix with rownames annotated to
Illumina array probe ID.}

\item{upstream}{An integer specifying upstream boundary of the promoter region
from transcription start site, the maximum is 5000, default for 1500.}

\item{downstream}{An integer specifying downstream boundary of the promoter region
from transcription start site, the maximum is 5000, default for 500.}

\item{method}{A parameter specifying how the methylation level for a particular
miRNA be calculated. "mean" for that average beta value of all CpGs within
promoter region, "nearest" for that the beta value of the nearest CpG from
the transcription start site as representative, "loess" for that the loess
fitted average beta value of all CpGs within promoter region.}
}
\value{
a vector of the methylation level for particular gene across all samples
}
\description{
Title Obtain beta value for each individual miRNA
}
\examples{
data(beta450k.m)
cpgPos181a5p.v <- c(-1105, -960, 63, 2362)
names(cpgPos181a5p.v) <- c("cg19639967", "cg07837852", "cg26257082", "cg09517106")
beta181a5p.v <- getGeneBeta(cpgPos181a5p.v, beta450k.m,  upstream=2000, downstream=2000, method="mean" )

}
