% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRExpRPM.m}
\alias{miRExpRPM.m}
\title{miRNA expression profile (TCGA BRCA)}
\format{
A matrix with 1349 rows (miRNAs) and 10 variables (samples)
}
\usage{
data(miRExpRPM.m)
}
\description{
An example data table that containing the miRNA expression RPM
values from 10 bulk tissue samples of breast cancer from TCGA (5 normal vs. 5
tumor). Only used for demonstration purpose.
}
\keyword{datasets}
