% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_scatter.R
\name{plot_t_scatter}
\alias{plot_t_scatter}
\title{Title Plot t-value scatter plot from differential promoter DNAm methylation
and differential expression.}
\usage{
plot_t_scatter(statM, statE, FDR_threshold = 0.05)
}
\arguments{
\item{statM}{matrix of differential methylation analysis for the miRNAs}

\item{statE}{matrix of differential expression analysis for the miRNAs}

\item{FDR_threshold}{FDR threshold for the significantly differentially methylated
miRNAs and differentially expressed miRNAs}
}
\value{
Graphical output of the t-values
}
\description{
Title Plot t-value scatter plot from differential promoter DNAm methylation
and differential expression.
}
\examples{
data(beta450k.m)
data(miRExpRPM.m)
miRBeta.m <- probe2miR(beta=beta450k.m, array="450k", type="Mature", upstream=2000, downstream=2000, method="mean")
pheno.v  <- c(rep("Normal", 5), rep("Tumor", 5))
statM.df <- DoWilcox(miRBeta.m, pheno.v, case = "Tumor", control = "Normal")
logExp.m <- log2(miRExpRPM.m + 1)
statE.df <- diffmiR(logExp.m, pheno.v, case = "Tumor", control = "Normal", method="limma")
plot_t_scatter(statM.df, statE.df, FDR_threshold = 0.05)

}
