% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe2miR.R
\name{probe2miR}
\alias{probe2miR}
\title{Title Obtain DNA methylation beta value matrix for all miRNAs}
\usage{
probe2miR(
  beta,
  array = "450k",
  IlmnID = FALSE,
  type = "Mature",
  upstream = 1500,
  downstream = 500,
  method = "mean"
)
}
\arguments{
\item{beta}{The DNA methylation beta value matrix with rownames annotated to
Illumina array probe IDs.}

\item{array}{array A parameter specifying the type of input data matrix, it should be
one of the "450k" for Illumina 450k array, or "EPIC" for Illumina EPIC (850k)
array, or "EPICv2" for Illumina EPICv2 (935k) array, default for "450k".}

\item{IlmnID}{A logical to indicate whether the IlmnID (which comprise the 'Name'
and a four-character suffix, e.g. 'cg09617579_BC12') was used for EPICv2 data, default
for FALSE.}

\item{type}{A parameter specifying the miRNA type used. it should be either
"mature" for mature miRNA, or "pre" for precursor miRNA, default for "mature".}

\item{upstream}{An integer specifying upstream boundary of the promoter region
from transcription start site, the maximum is 5000, default for 1500.}

\item{downstream}{An integer specifying downstream boundary of the promoter region
from transcription start site, the maximum is 5000, default for 500.}

\item{method}{A parameter specifying how the methylation level for a particular
miRNA be calculated. "mean" for that average beta value of all CpGs within
promoter region, "nearest" for that the beta value of the nearest CpG from
the transcription start site as representative, "loess" for taht the loess
fitted average beta value of all CpGs within promoter region.}
}
\value{
DNA methylation matrix with rows are miRNAs columns are samples
}
\description{
Title Obtain DNA methylation beta value matrix for all miRNAs
}
\examples{
data(beta450k.m)
miRBeta.m <- probe2miR(beta=beta450k.m, array="450k", type="Mature", upstream=2000, downstream=2000, method="mean")

}
